Modified function for converting. Updating tests and file structure.
