+++ src/packages/eightballer/connections/ocean/connection.yaml
@@ -8,7 +8,7 @@ license: Apache-2.0
 aea_version: '>=1.0.0, <2.0.0'
 fingerprint:
   __init__.py: QmZvYZ5ECcWwqiNGh8qNTg735wu51HqaLxTSifUxkQ4KGj
-  connection.py: Qmbt3VU5QUT5v734H9SzD43nShQ99wxt7yYjS1iqXy9bFC
+  connection.py: QmPZXbRahRkKMF9fBWezF6HxaRm7ncVqz2vRZF2T88aFb4
   readme.md: Qmdt71SaCCwAG1c24VktXDm4pxgUBiPMg4bWfUTiqorypf
 fingerprint_ignore_patterns: []
 connections: []
